% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_uniprot_proteomes.R
\name{cur_uniprot_proteomes}
\alias{cur_uniprot_proteomes}
\title{Get current available UniProt reference proteomes}
\usage{
cur_uniprot_proteomes()
}
\value{
Returns a \code{data.frame} with the current species in UniProt with
available reference proteomes.
}
\description{
This function outputs a table showing which UniProt reference
proteomes are available to download from their
\href{https://beta.uniprot.org/help/downloads}{FTP site}. This table is updated
by UniProt every 8 weeks.
}
\details{
Some proteomes in UniProt have been (manually and algorithmically)
selected as reference proteomes. They cover well-studied model organisms and
other organisms of interest for biomedical research and phylogeny. For all
of the \code{proteome_id} listed in this table you can download the following types
of files:
\itemize{
\item \verb{*.gene2acc}, \verb{*.fasta}, \verb{*.dat}, \verb{*.xml} (canonical sequences)
\item \verb{*_DNA.fasta} (only canonical sequences)
\item \verb{*.idmapping}
}

For a subset of the \code{proteome_id} listed in this table you can also download:
\itemize{
\item \verb{*_additional.fasta}, \verb{*_additional.dat}, \verb{*_additional.xml} (isoform sequences)
\item \verb{*_DNA.miss} (only canonical sequences)
}
}
\examples{
# To do: example

}
