% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_uniprot_release.R
\name{cur_uniprot_release}
\alias{cur_uniprot_release}
\title{Get current UniProt release number}
\usage{
cur_uniprot_release()
}
\value{
Returns a string with the current UniProt release number in the
format \code{"20XX_0X"}.
}
\description{
Get current UniProt release number
}
\examples{
# Useful for naming files e.g.
\dontrun{
  Biostrings::writeXStringSet(
    x = Biostrings::AAStringSet("AAA"),
    filepath = paste0(cur_uniprot_release, "_a_protein.fasta")
  )
}

}
