% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\docType{data}
\name{formats}
\alias{formats}
\title{(Dataset) UniProt download formats}
\format{
An object of class \code{data.frame} with 144 rows and 3 columns.
}
\source{
UniProtKB download formats have been determined by hand by querying
each REST API database with an unavailable format (usually \code{txt}) and
determining the allowed formats from the resulting error response.
}
\usage{
formats
}
\description{
This dataframe contains the available download file types i.e.
values for \code{format} for each UniProt database.
See \strong{Examples} below for how to use this object.
}
\details{
Columns:
\tabular{ll}{
\strong{func} \tab \code{character}. \code{uniprotREST} function to be used. \cr
\strong{database} \tab \code{character}. UniProt database to be queried. \cr
\strong{format} \tab \code{character}. The different formats available to download
i.e. the string you'll use with the \code{uniprotREST} function. \cr
}
}
\examples{
# What UniProt databases are available to query?
levels(formats$database)

# What formats are available for querying the `proteomes` database,
# using `uniprot_search()`
formats[formats$database == "proteomes" & formats$func == "search", "format"]
}
\keyword{datasets}
