% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to.R
\docType{data}
\name{from_to}
\alias{from_to}
\title{Available ID mapping databases}
\format{
An object of class \code{data.frame} with 98 rows and 4 columns.
}
\source{
\href{https://www.uniprot.org/help/id_mapping}{UniProt ID mapping}
}
\usage{
from_to
}
\description{
A dataset containing all the databases available to use with
\code{uniprot_map()}. Generally, any \code{from} can be mapped \code{to = "UniProtKB"} or
\code{"UniProtKB-Swiss-Prot"} and \code{from = "UniProtKB_AC-ID"} can be mapped to any
\code{to}.

See the \code{from_to} object for a list of allowed \code{from}-\code{to} mappings.
Alternatively use the function \code{\link{check_from_to}} to check if
the \code{from} and \code{to} pair is valid.
}
\details{
Columns:
\tabular{ll}{
\strong{name} \tab \code{character}. Database name. \cr
\strong{from} \tab \code{logical}. Can the IDs be mapped from? \cr
\strong{to} \tab \code{logical}. Can the IDs be mapped to? \cr
\strong{url} \tab \code{character}. URLs for each database (replace \%id with a given ID) \cr
}
}
\keyword{datasets}
