% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to_dbs.R
\docType{data}
\name{from_to_dbs}
\alias{from_to_dbs}
\title{(Dataset) From/to databases for ID mapping}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\source{
From/to pairs have been downloaded according to the "Valid from and
to databases pairs" section on the UniProt
\href{https://www.uniprot.org/help/id_mapping}{ID Mapping} page.
}
\usage{
from_to_dbs
}
\description{
This dataframe contains details of the databases that can be
mapped from/to with \code{\link[=uniprot_map]{uniprot_map()}}. See \link{from_to_rules} for the rules on
which database identifiers can be mapped to what other databases.
}
\details{
Columns:
\tabular{ll}{
\strong{name} \tab \code{character}, database name. \cr
\strong{from} \tab \code{character}, is the database allowed to be mapped \code{from}? \cr
\strong{to} \tab \code{character}, is the database allowed to be mapped \code{to}? \cr
\strong{url} \tab \code{character}, database URL. \cr
\strong{formats_db} \tab \code{factor}, relevant \code{formats} database, see \link{formats}
for which download formats are available \cr
}
}
\seealso{
\link{from_to_rules}
}
\keyword{datasets}
