% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to_rules.R
\docType{data}
\name{from_to_rules}
\alias{from_to_rules}
\title{UniProt ID mapping from/to rules}
\format{
An object of class \code{list} of length 98.
}
\source{
From/to pairs have been downloaded according to the "Valid from and
to databases pairs" section on the UniProt
\href{https://www.uniprot.org/help/id_mapping}{ID Mapping} page.
}
\usage{
from_to_rules
}
\description{
This list contains the valid from/to pairings that can be used
with \code{uniprot_map()}.

\code{names(from_to_rules)} shows all the valid values for \code{from}.

\code{from_to_rules[["SGD"]]} shows all the valid \code{to} values for a certain
\code{from}, for example in this case is \code{"SGD"}.
}
\seealso{
\link{from_to_dbs} for details of each database available for ID mapping.
}
\keyword{datasets}
