% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uniprot_proteome.R
\name{get_uniprot_proteome}
\alias{get_uniprot_proteome}
\title{Download canonical reference proteome from UniProt FTP server}
\usage{
get_uniprot_proteome(
  id,
  format = c("fasta", "dat", "xml", "gene2acc", "idmapping", "DNA.fasta", "DNA.miss"),
  path = NULL,
  isoform = FALSE,
  dry_run = FALSE
)
}
\arguments{
\item{id}{\code{string}, UniProt proteome ID e.g. \code{UP000005640} for \emph{Homo sapiens}.}

\item{format}{\code{string}, desired file format to download. Choose from:
\code{fasta} (default), \code{dat}, \code{xml}, \code{gene2acc}, \code{idmapping}, \code{DNA.fasta}, \code{DNA.miss}.
See \href{https://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/reference_proteomes/README}{here}
for a description of each file type.}

\item{path}{Optional \code{string}. Path to save the downloaded file.}

\item{isoform}{\code{logical}, should the file with isoform sequences be downloaded?
Only applicable for the \code{fasta}, \code{dat}, and \code{xml} formats. Default is \code{FALSE}.}

\item{dry_run}{\code{logical}, default is \code{FALSE}. If \code{TRUE} print the HTTP
request using \code{\link[httr2]{req_dry_run}} without actually sending
anything to the UniProt server. Useful for debugging if you get an HTTP 400
Bad Request error i.e. \code{id} was not written correctly, or if you get an
HTTP 404 Not found error i.e. the desired \code{format} is not available for your
proteome of interest.}
}
\value{
If the request was successful (i.e. the request was successfully
performed and a response with HTTP status code <400 was received), an HTTP
response is returned; otherwise an error is thrown. Use the \code{path} argument
to save the body of the response to a file.

If \code{dry_run = TRUE}, then prints the HTTP request and returns,
invisibly, a list containing information about the request without actually
sending anything to the UniProt server.
}
\description{
To do: write proper description. See
\href{https://pwilmart.github.io/blog/2020/09/19/shotgun-quantification-part2}{here}
for description of the FTP canonical reference proteomes.
}
\examples{

\dontrun{
get_uniprot_proteome("UP000005640", "fasta", "./Homo_sapiens_canonical.fasta")
}

}
