% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_from_to.R
\docType{data}
\name{map_from_to}
\alias{map_from_to}
\title{ID mapping from/to pairs}
\format{
An object of class \code{list} of length 97.
}
\source{
\href{https://www.uniprot.org/help/id_mapping}{UniProt ID mapping}
}
\usage{
map_from_to
}
\description{
Description.
}
\details{
Details.
}
\keyword{datasets}
