% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match.arg.exact}
\alias{match.arg.exact}
\title{Argument verification using exact matching}
\source{
This code was copied from the source of the
\href{https://CRAN.R-project.org/package=MARSS}{MARSS} package which is licensed
under CC0 1.0 Universal.
}
\usage{
match.arg.exact(arg, choices, several.ok = FALSE, exact = TRUE)
}
\arguments{
\item{arg}{\verb{character vector} (of length one unless \code{several.ok = TRUE}) or
\code{NULL}.}

\item{choices}{\verb{character vector} of candidate values.}

\item{several.ok}{\code{logical} specifying if \code{arg} should be allowed to have
more than one element. Default is \code{FALSE}.}

\item{exact}{\code{logical} specifying if exact matching should be used. Default
is \code{TRUE}.}
}
\value{
Returns the exact match if there is one, otherwise an returns an
error if \code{several.ok = FALSE}. When \code{several.ok = TRUE} and more than one
element of \code{arg} has a match, all matches are returned.
}
\description{
Works the same as \code{\link[base]{match.arg}} but
with the default of using exact matching with \code{\link[base]{match}}
instead of partial matching with \code{\link[base]{pmatch}}.
}
