% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteotools-package.R
\docType{package}
\name{proteotools-package}
\alias{proteotools}
\alias{proteotools-package}
\title{proteotools: General functions for proteomics}
\description{
Provides simple functions to facilitate proteomics analyses. Should be used alongside other proteomics and mass spectrometry focused packages.
}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID})

}
\keyword{internal}
