% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_map.R
\name{uniprot_map}
\alias{uniprot_map}
\title{Map from/to UniProt IDs}
\usage{
uniprot_map(
  ids,
  from = "UniProtKB_AC-ID",
  to = "UniProtKB",
  format = "tsv",
  path = NULL,
  fields = NULL,
  isoform = NULL,
  method = "paged",
  page_size = 500,
  compressed = NULL,
  verbosity = NULL,
  dry_run = FALSE
)
}
\arguments{
\item{ids}{\code{character}, vector of identifiers to map from.}

\item{from}{\code{string}, database to map from. Default is \code{"UniProtKB_AC-ID"}.
See \link{from_to_dbs} possible databases whose identifiers you can map from.}

\item{to}{\code{string}, database to map to. Default is \code{"UniProtKB"}.
See \link{from_to_rules} for the possible databases you can map to, depending on
the \code{from} database.}

\item{format}{\code{string}, data format to fetch. Default is \code{"tsv"}.
Can only be \code{"tsv"} at the moment.}

\item{path}{Optional \code{string}, file path to save the results, e.g.
\code{"path/to/results.tsv"}.}

\item{fields}{Optional \code{character}, fields (i.e. columns) of data to get.
Only used if \code{to} is a UniProtKB, UniRef, or UniParc database. See
\link{return_fields} for all available fields.}

\item{isoform}{Optional \code{logical}, should protein isoforms be included in the
results? Not necessarily relevant for all formats and databases.}

\item{method}{\code{string}, download method to use. Either \code{"paged"} (default) or
\code{"stream"}. Paged is more robust to connection issues and takes less
memory. Stream may be faster, but uses more memory and is more sensitive
to connection issues.}

\item{page_size}{Optional \code{integer}, how many entries per page to request?
Only relevant if \code{method = "paged"}. It's best to leave this at \code{500}.}

\item{compressed}{Optional \code{logical}, should gzipped data be requested?
Only relevant if \code{method = "stream"} and \code{path} is specified.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{req_verbose()} that uses an integer to control verbosity:
\itemize{
\item 0: no output
\item 1: show headers
\item 2: show headers and bodies
\item 3: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}

\item{dry_run}{\code{logical}, perform request with \code{\link[httr2:req_dry_run]{httr2::req_dry_run()}}?
Default is \code{FALSE}.}
}
\value{
By default, returns an object whose type depends on \code{format}:
\itemize{
\item \strong{\code{tsv}}: \code{data.frame}
}

If \code{path} is specified, saves the results to the file path indicated,
and returns \code{NULL} invisibly. If \code{dry_run = TRUE}, returns a
list containing information about the request, including the request
\code{method}, \code{path}, and \code{headers}.
}
\description{
This function wraps the
\href{https://www.uniprot.org/help/id_mapping}{UniProt ID Mapping} service which
maps between the identifiers used in one database, to the identifiers of
another. By default it maps UniProtKB accessions to UniProt and returns a
\code{data.frame}, so you can get information about a set of IDs e.g. gene names,
protein lengths, domains, etc. You can also map IDs from/to other databases
e.g. \verb{from = "Ensembl", to = "UniProtKB"}.
}
\examples{
\dontrun{
  # Default, get info about UniProt IDs
  uniprot_map(
    "P99999",
    format = "tsv",
    fields = c("accession", "gene_primary", "feature_count")
  )

  # Other common use, mapping other IDs to UniProt
  # (or vice-versa)
  uniprot_map(
    c("ENSG00000088247", "ENSG00000162613"),
    from = "Ensembl",
    to = "UniProtKB"
  )

}
}
\seealso{
Other API wrapper functions: \code{\link[=uniprot_search]{uniprot_search()}}, \code{\link[=uniprot_single]{uniprot_single()}}
}
