% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_request.R
\name{uniprot_request}
\alias{uniprot_request}
\title{Create UniProt HTTP request}
\usage{
uniprot_request(url, method = "GET", ..., max_tries = 5, rate = 1/1)
}
\arguments{
\item{url}{\code{string}, the URL to make the request.}

\item{method}{\code{string}, the HTTP request method. One of \code{"GET"} (default),
\code{"POST"}, or \code{"HEAD"}.}

\item{...}{Name-value pairs that provide query parameters. Each value must be
either a length-1 atomic vector (which is automatically escaped) or \code{NULL}
(which is silently dropped).}

\item{max_tries}{\code{integer}, the number of maximum attempts to perform the
HTTP request. Default is \code{5}.}

\item{rate}{\code{numeric}, the maximum number of requests per second. Default is
\code{1 / 1} i.e. 1 request per 1 second.}
}
\value{
Returns an \code{httr2_request} object.
}
\description{
The function creates a request object for the UniProt REST API.
You likely won't use this function directly, but rather one of the
wrapper functions: \code{\link[=uniprot_map]{uniprot_map()}}, \code{\link[=uniprot_search]{uniprot_search()}}, or \code{\link[=uniprot_single]{uniprot_single()}}.
}
\examples{
# Construct a request for the accession and
# gene name for human Cytochrome C
uniprot_request(
  url = "https://rest.uniprot.org/uniprotkb/P99999",
  fields = "accession,gene_primary",
  format = "tsv"
)

}
