% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_retrieve.R
\name{uniprot_retrieve}
\alias{uniprot_retrieve}
\title{Retrieve individual UniProt entry}
\usage{
uniprot_retrieve(
  id,
  database = c("uniprotkb", "uniref", "uniparc", "proteomes", "taxonomy", "keywords",
    "citations", "diseases", "database", "locations", "unirule", "arba"),
  format = c("tsv", "fasta", "json"),
  path = NULL,
  fields = NULL,
  isoform = NULL,
  verbosity = NULL,
  dry_run = FALSE
)
}
\arguments{
\item{id}{\code{string}, the entry's unique identifier.}

\item{database}{\code{string}, the name of the database endpoint of interest.
Choose from: \code{uniprotkb} (default), \code{uniref}, \code{uniparc}, \code{proteomes}, \code{taxonomy},
\code{keywords}, \code{citations}, \code{diseases}, \code{database}, \code{locations}, \code{unirule}, \code{arba}.
See details below.}

\item{format}{\code{string}, output format for search results.
Only \code{tsv} (default) or \code{json} implemented at the moment.}

\item{path}{Optional \code{string}, file path to save API response. Useful for
large queries to avoid storing in memory.}

\item{fields}{Optional \verb{character vector}. Columns to retrieve in the search
results. Applies to the \code{json}, \code{tsv}, and \code{xlsx} formats only.
See \href{https://www.uniprot.org/help/return_fields}{here} for available
UniProtKB fields.}

\item{isoform}{Optional \code{logical}. Whether or not to include isoforms in the
search results. Only applicable if \code{database = "uniprotkb"}.}

\item{verbosity}{Optional \code{integer}, how much information to print? This is
a wrapper around \code{\link[httr2]{req_verbose}} that uses an integer to
control verbosity:
\itemize{
\item 0: no output.
\item 1: show headers.
\item 2: show headers and bodies.
\item 3: show headers, bodies, and curl status messages.
}}

\item{dry_run}{\code{logical}, default is \code{FALSE}. If \code{TRUE} print the HTTP
request using \code{\link[httr2]{req_dry_run}} without actually sending
anything to the UniProt server. Useful for debugging if you get an HTTP 400
Bad Request error i.e. your \code{query} was not written correctly.}
}
\value{
If the request was successful (i.e. the request was successfully
performed and a response with HTTP status code <400 was received), a data.frame
(\code{tsv}) or a list (\code{json}) is returned; otherwise an error is thrown.
Use the \code{path} argument to save to a file instead.

If \code{dry_run = TRUE}, then prints the HTTP request and returns,
invisibly, a list containing information about the request without actually
sending anything to the UniProt server.
}
\description{
To do: describe.
}
\examples{
# To do: add examples

}
