% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cur_uniprot_release.R
\name{cur_uniprot_release}
\alias{cur_uniprot_release}
\title{Get current UniProt release number}
\usage{
cur_uniprot_release()
}
\value{
Returns a \code{string} with the current UniProt release number in the
format \code{"20XX_0X"}.
}
\description{
This function queries the UniProt REST API and gets the current
release number.
}
\examples{
# Useful for constructing file names e.g.
\dontrun{
paste0(cur_uniprot_release(), "_a_protein.fasta")
}

}
