% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_release.R
\name{current_release}
\alias{current_release}
\title{Get current UniProt release number}
\usage{
current_release(dry_run = FALSE)
}
\arguments{
\item{dry_run}{\code{logical}, default is \code{FALSE}. Used for debugging purposes.
If \code{TRUE} print the HTTP
request using \code{\link[httr2]{req_dry_run}} without actually sending
anything to the UniProt server.}
}
\value{
Returns a \code{string} with the current UniProt release number in the
format \code{"DDDD_DD"} e.g. \verb{2022_02}.
}
\description{
This function queries the UniProt REST API and gets the current
release number.
}
\examples{
# Useful for constructing file names e.g.
\dontrun{
paste0(current_release(), "_a_protein.fasta")
}

}
