% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_paged.R
\name{fetch_paged}
\alias{fetch_paged}
\title{Fetch results via pagination}
\usage{
fetch_paged(req, n_pages, format = "tsv", path = NULL, verbosity = NULL)
}
\arguments{
\item{req}{\code{httr2_request} object, generated by e.g. \code{\link[httr2:request]{httr2::request()}} or
\code{\link[=uniprot_request]{uniprot_request()}}.}

\item{n_pages}{\code{integer}, the number of pages to be fetched. This can be
calculated by dividing the number of total results by the page size e.g.
\code{resp$headers$x-total-results / page_size}.}

\item{format}{\code{string}, data format to fetch. Can only be \code{"tsv"} at the
moment.}

\item{path}{path Optional \code{string}, file path to save the results, e.g.
\code{"path/to/results.tsv"}.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{req_verbose()} that uses an integer to control verbosity:
\itemize{
\item 0: no output
\item 1: show headers
\item 2: show headers and bodies
\item 3: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}
}
\value{
By default, returns an object whose type depends on \code{format}:
\itemize{
\item \strong{tsv}: \code{data.frame}
}

If \code{parse = FALSE}, returns an \code{httr2_response}. If \code{path} is specified,
saves the parsed results to the file path indicated, and returns \code{NULL}
invisibly.
}
\description{
This function performs a request for data from the UniProt REST
API, fetches the results using pagination, and saves them to a file or
into memory.

You likely won't use this function directly, but rather one of the
wrapper functions: \code{\link[=uniprot_map]{uniprot_map()}}, \code{\link[=uniprot_search]{uniprot_search()}}, or \code{\link[=uniprot_single]{uniprot_single()}}.
\subsection{Things to note:}{
\enumerate{
\item The pagination input is less expensive for the API infrastructure to
deal with as the memory demand is distributed over a longer period of time.
\item If the connection is interrupted while fetching results with pagination,
only the request for the current page needs to be reattempted. With the
stream endpoint, the entire request needs to be completely restarted.
}
}
}
\examples{
\dontrun{
  req <- uniprot_request(
    "https://rest.uniprot.org/uniref/search",
    query = "P99999",
    format = "tsv",
    fields = "id,name,count",
    size = 1
  )

  fetch_paged(req, n_pages = 3)
}
}
