% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to.R
\docType{data}
\name{from_to}
\alias{from_to}
\title{Available ID mapping databases}
\format{
An object of class \code{data.frame} with 98 rows and 4 columns.
}
\source{
\href{https://www.uniprot.org/help/id_mapping}{UniProt ID mapping}
}
\usage{
from_to
}
\description{
A dataset containing all the databases available to use with
\code{uniprot_map()}.
}
\details{
Columns:
\tabular{ll}{
\strong{name} \tab \code{character}. Database name. \cr
\strong{from} \tab \code{logical}. Can the IDs be mapped from? \cr
\strong{to} \tab \code{logical}. Can the IDs be mapped to? \cr
\strong{url} \tab \code{character}. URLs for each database (replace \%id with a given ID) \cr
}
}
\keyword{datasets}
