% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to_dbs.R
\docType{data}
\name{from_to_dbs}
\alias{from_to_dbs}
\title{UniProt ID mapping from/to databases}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\source{
From/to pairs have been downloaded according to the "Valid from and
to databases pairs" section on the UniProt
\href{https://www.uniprot.org/help/id_mapping}{ID Mapping} page.
}
\usage{
from_to_dbs
}
\description{
This dataframe contains details of the databases that can be
mapped from/to with \code{uniprot_map()}.
}
\details{
Columns:
\tabular{ll}{
\strong{name} \tab \code{character}. Database name. \cr
\strong{from} \tab \code{character}. Is the database allowed to be mapped \code{from}? \cr
\strong{to} \tab \code{character}. Is the database allowed to be mapped \code{to}? \cr
\strong{url} \tab \code{character}. Database URL. \cr
\strong{return_fields_db} \tab \code{character}. Relevant return fields database,
see \link{return_fields} for which fields i.e. columns can be requested. \cr
}
}
\seealso{
\link{from_to_rules} for rules of which database identifiers can be
mapped to what other database.
}
\keyword{datasets}
