% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_to_rules.R
\docType{data}
\name{from_to_rules}
\alias{from_to_rules}
\title{(Dataset) From/to rules for ID mapping}
\format{
An object of class \code{list} of length 98.
}
\source{
From/to pairs have been downloaded according to the "Valid from and
to databases pairs" section on the UniProt
\href{https://www.uniprot.org/help/id_mapping}{ID Mapping} page.
}
\usage{
from_to_rules
}
\description{
This list contains the valid from/to pairings that can be used
with \code{\link[=uniprot_map]{uniprot_map()}}. See \strong{Examples} below for how to use this object.
Also see \link{from_to_dbs} for the more information on each database.
}
\examples{
# Show valid `from` values
names(from_to_rules)

# Show valid `to` values for a given `from` e.g. SGD
from_to_rules[["SGD"]]
}
\seealso{
\link{from_to_dbs}
}
\keyword{datasets}
