% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_fields.R
\docType{data}
\name{return_fields}
\alias{return_fields}
\title{UniProt return fields}
\format{
An object of class \code{data.frame} with 411 rows and 4 columns.
}
\source{
UniProtKB return fields have been scraped from the
\href{https://www.uniprot.org/help/return_fields}{UniProtKB return fields} page.
The return fields from other Uniprot databases have been determined by hand
using Web Developer Tools (F12) to inspect the GET request made when
searching the different database on the UniProt website.
}
\usage{
return_fields
}
\description{
This dataframe contains the valid fields (i.e. columns) of
data you can request from UniProt. The strings in the \code{field} column is what
you'll use with the functions in this package.
}
\details{
Columns:
\tabular{ll}{
\strong{database} \tab \code{character}. UniProt database to be queried. \cr
\strong{section} \tab \code{character}. Similar return fields are grouped together in sections. \cr
\strong{field} \tab \code{character}. The return field i.e. the string used to request the desired column of information. \cr
\strong{label} \tab \code{character}. Human readable column name that will be returned by the API.\cr
}
}
\keyword{datasets}
