% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_map.R
\name{uniprot_map}
\alias{uniprot_map}
\title{Map from/to UniProt IDs}
\usage{
uniprot_map(
  ids,
  from = "UniProtKB_AC-ID",
  to = "UniProtKB",
  format = "tsv",
  path = NULL,
  fields = NULL,
  isoform = NULL,
  compressed = FALSE,
  verbosity = NULL,
  dry_run = FALSE
)
}
\arguments{
\item{ids}{\code{string}, comma-separated IDs to map.}

\item{from}{\code{string}, name of database which IDs are from. The default is
\code{UniProtKB_AC-ID}. See \code{\link{from_to}} for a list of all possible
database names.}

\item{to}{\code{string}, name of database to map to. The default is \code{UniProtKB}.
See \code{\link{from_to}} for a list of all possible database names.}

\item{format}{\code{string}, output format for mapping results.
To do: add available formats.}

\item{path}{Optional \code{string}. Path to save the body of the HTTP response.
This is useful for large search results since it avoids storing the response
in memory.}

\item{fields}{Optional \verb{character vector}. Columns to retrieve in the search
results. Applies to the \code{tsv}, and \code{xlsx} formats only.}

\item{isoform}{Optional \code{logical}. Whether or not to include isoforms in the
search results. Only applicable if \code{to} is \code{UniProtKB} or \code{UniProt-Swiss-Prot}.}

\item{compressed}{\code{logical}, should results be returned gzipped? Default is
\code{FALSE}.}

\item{verbosity}{Optional \code{integer}, how much information to print? This is
a wrapper around \code{\link[httr2]{req_verbose}} that uses an integer to
control verbosity:
\itemize{
\item 0: no output.
\item 1: show headers.
\item 2: show headers and bodies.
\item 3: show headers, bodies, and curl status messages.
}}

\item{dry_run}{\code{logical}, default is \code{FALSE}. If \code{TRUE} print the HTTP
request using \code{\link[httr2]{req_dry_run}} without actually sending
anything to the UniProt server. Useful for debugging if you get an HTTP 400
Bad Request error i.e. your \code{ids} or \code{fields} were not written correctly.}
}
\value{
If the request was successful (i.e. the request was successfully
performed and a response with HTTP status code <400 was received), an HTTP
response is returned; otherwise an error is thrown. Use the \code{path} argument
to save the body of the response to a file.

If \code{dry_run = TRUE}, then prints the HTTP request and returns,
invisibly, a list containing information about the request without actually
sending anything to the UniProt server.
}
\description{
To do: write description.
}
\examples{
# To do: add some examples

}
