% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_search.R
\name{uniprot_search}
\alias{uniprot_search}
\title{Search UniProt}
\usage{
uniprot_search(
  query,
  database = "uniprotkb",
  format = "tsv",
  path = NULL,
  fields = NULL,
  isoform = NULL,
  method = "paged",
  page_size = 500,
  compressed = NULL,
  verbosity = NULL,
  dry_run = FALSE
)
}
\arguments{
\item{query}{\code{string}, the search query. See
\href{https://www.uniprot.org/help/text-search}{this page} for helping
constructing search queries.}

\item{database}{\code{string}, database to look up. Default is \code{"uniprotkb"}.
See the \strong{Databases} section below for all available databases.}

\item{format}{\code{string}, data format to fetch. Default is \code{"tsv"}.
Can only be \code{"tsv"} at the moment.}

\item{path}{Optional \code{string}, file path to save the results, e.g.
\code{"path/to/results.tsv"}.}

\item{fields}{Optional \code{character}, fields (i.e. columns) of data to get.
The fields available depends on the database used, see \link{return_fields}
for all available fields.}

\item{isoform}{Optional \code{logical}, should protein isoforms be included in the
results? Not necessarily relevant for all formats and databases.}

\item{method}{\code{string}, download method to use. Either \code{"paged"} (default) or
\code{"stream"}. Paged is more robust to connection issues and takes less
memory. Stream may be faster, but uses more memory and is more sensitive
to connection issues.}

\item{page_size}{Optional \code{integer}, how many entries per page to request?
Only relevant if \code{method = "paged"}. It's best to leave this at \code{500}.}

\item{compressed}{Optional \code{logical}, should gzipped data be requested?
Only relevant if \code{method = "stream"} and \code{path} is specified.}

\item{verbosity}{How much information to print? This is a wrapper
around \code{req_verbose()} that uses an integer to control verbosity:
\itemize{
\item 0: no output
\item 1: show headers
\item 2: show headers and bodies
\item 3: show headers, bodies, and curl status messages.
}

Use \code{\link[httr2:with_verbosity]{with_verbosity()}} to control the verbosity of requests that
you can't affect directly.}

\item{dry_run}{\code{logical}, perform request with \code{\link[httr2:req_dry_run]{httr2::req_dry_run()}}?
Requires the \code{httpuv} package to be installed. Default is \code{FALSE}.}
}
\value{
By default, returns an object whose type depends on \code{format}:
\itemize{
\item \strong{\code{tsv}}: \code{data.frame}
}

If \code{path} is specified, saves the results to the file path indicated,
and returns \code{NULL} invisibly. If \code{dry_run = TRUE}, returns a
list containing information about the request, including the request
\code{method}, \code{path}, and \code{headers}.
}
\description{
Search Uniprot via REST API. By default it searches the supplied
query against UniProtKB and returns a \code{data.frame} of matching proteins.
It is a wrapper for
\href{https://www.uniprot.org/help/api_queries}{this} UniProt API
endpoint.
}
\section{Databases}{


The following databases are available to query:
\itemize{
\item \strong{\code{uniprotkb}}: \href{https://www.uniprot.org/help/uniprotkb}{UniProt Knowledge Base}
\item \strong{\code{uniref}}: \href{https://www.uniprot.org/help/uniref}{UniProt Reference Clusters}
\item \strong{\code{uniparc}}: \href{https://www.uniprot.org/help/uniparc}{UniProt Archive}
\item \strong{\code{proteomes}}: \href{https://www.uniprot.org/help/reference_proteome}{Reference proteomes}
\item \strong{\code{taxonomy}}: \href{https://www.uniprot.org/help/taxonomy}{Taxonomy}
\item \strong{\code{keywords}}: \href{https://www.uniprot.org/help/keywords}{Keywords}
\item \strong{\code{citations}}: \href{https://www.uniprot.org/help/literature_references}{Literature references}
\item \strong{\code{diseases}}: \href{https://www.uniprot.org/help/disease_query}{Disease queries}
\item \strong{\code{database}}: \href{https://www.uniprot.org/help/cross_references_section}{Cross references}
\item \strong{\code{locations}}: \href{https://www.uniprot.org/help/subcellular_location}{Subcellular location}
\item \strong{\code{unirule}}: \href{https://www.uniprot.org/help/unirule}{UniRule}
\item \strong{\code{arba}}: \href{https://www.uniprot.org/help/automatic_annotation}{ARBA}
(Association-Rule-Based Annotator)
}
}

\examples{
\dontrun{
  # Search for all human glycoproteins from SwissProt
  res <- uniprot_search(
    query = "(proteome:UP000005640) AND (keyword:KW-0325) AND (reviewed:true)",
    database = "uniprotkb",
    format = "tsv",
    fields = c("accession", "gene_primary", "feature_count")
  )

  # Look at the resulting dataframe
  head(res)
}
}
\seealso{
Other API wrapper functions: \code{\link[=uniprot_map]{uniprot_map()}}, \code{\link[=uniprot_single]{uniprot_single()}}
}
